+++ src/Data/SExpresso/Parse/Generic.hs
@@ -24,12 +24,6 @@
 
 module Data.SExpresso.Parse.Generic
   (
-    SListDefinition(..),
-    getAtomParser,
-    setAtomParser,
-    getSepRule,
-    setSepRule,
-    
     SeparatorRule(..),
     sepIsMandatory,
     sepIsOptional,
@@ -39,49 +33,21 @@ module Data.SExpresso.Parse.Generic
     sepByRule1,
     sepEndByRule,
     sepEndByRule1,
+
     sepByList,
+    sepByList1,
+    sepEndByList,
+    sepEndByList1,
     
-    parseSExpr,
-    decodeOne,
-    decode
+    sexpr,
+    manySExpr,
+    manySExpr1
    )
   where
 
 import Control.Applicative
 import Text.Megaparsec
 import Data.SExpresso.SExpr
-
--- | The @'SExprParser' m b a@ datatype defines how to parse an
--- @'SExpr' b a@. Most parsing functions require the underlying monad
--- @m@ to be an instance of ('MonadParsec' e s m).
-data SListDefinition m a = forall b c d .  
-  -- | The @b@ @c@ and @d@ are there only for conviniences. There values are never used, but it is simpler to use than ().
-  SListDefinition 
-  (m b) -- ^ The parser for the opening tag. Returns an object of an
-        -- arbitrary type @c@ that will be used to create the closing
-        -- tag parser.
-  (m c) -- ^ A function that takes the object returned by the
-             -- opening tag parser and provide a parser for the
-             -- closing tag.
-  (m a) -- ^ The parser for atoms
-  (m d) -- ^ A parser for space tokens which does not accept empty
-         -- input (e.g. 'Text.Megaparsec.Char.space1')
-  (a -> a -> SeparatorRule) -- ^ A function to tell if two consecutive
-  -- atoms must be separated by space or
-  -- not. See also 'mkSpacingRule' and
-  -- 'setSepRule'
-
-getAtomParser :: SListDefinition m a -> m a
-getAtomParser (SListDefinition _ _ a _ _) = a
-
-getSepRule :: SListDefinition m a -> (a -> a -> SeparatorRule)
-getSepRule (SListDefinition _ _ _ _ r) = r
-
-setAtomParser :: m b -> (b -> b -> SeparatorRule) -> SListDefinition m a -> SListDefinition m b
-setAtomParser a r (SListDefinition s e _ sp _) = SListDefinition s e a sp r
-
-setSepRule :: (a -> a -> SeparatorRule) -> SListDefinition m a -> SListDefinition m a
-setSepRule r (SListDefinition s e a sp _) = SListDefinition s e a sp r
   
 -- | The 'SeparatorRule' datatype indicates if a separator is optional or mandatory between two consecutive tokens.
 data SeparatorRule =
@@ -206,44 +172,91 @@ sepEndByRule1 p sep rule = p >>= go
         selectFailure (TrivialError _ unexpected2 expected) = failure unexpected2 expected
         selectFailure (FancyError _ expected) = fancyFailure expected
 
--- | The 'parseSExprList' function return a parser for parsing S-expression of the form @'SList' _ _@.
+-- | The 'sepByList' function return a parser for parsing S-expression of the form @'SList' _ _@.
 sepByList :: (MonadParsec e s m) =>
-              SListDefinition m a -> 
-              m [a]
-sepByList (SListDefinition pSTag pETag atom sep sepRule)  =
- pSTag *> optional sep *> sepEndByRule atom sep sepRule <* pETag
+             m b ->
+             m c ->
+             m a ->
+             m sep ->
+             (a -> a -> SeparatorRule) ->
+             m [a]
+sepByList s e a sep r= between s e (sepByRule a sep r)
+
+
+-- | The 'sepByList' function return a parser for parsing S-expression of the form @'SList' _ _@.
+sepByList1 :: (MonadParsec e s m) =>
+             m b ->
+             m c ->
+             m a ->
+             m sep ->
+             (a -> a -> SeparatorRule) ->
+             m [a]
+sepByList1 s e a sep r= between s e (sepByRule1 a sep r)
+
+-- | The 'sepEndByList' function return a parser for parsing S-expression of the form @'SList' _ _@.
+sepEndByList :: (MonadParsec e s m) =>
+             m b ->
+             m c ->
+             m a ->
+             m sep ->
+             (a -> a -> SeparatorRule) ->
+             m [a]
+sepEndByList s e a sep r  = between s e (sepEndByRule a sep r)
+
+
+-- | The 'sepEndByList' function return a parser for parsing S-expression of the form @'SList' _ _@.
+sepEndByList1 :: (MonadParsec e s m) =>
+             m b ->
+             m c ->
+             m a ->
+             m sep ->
+             (a -> a -> SeparatorRule) ->
+             m [a]
+sepEndByList1 s e a sep r  = between s e (sepEndByRule1 a sep r)
+
+sexprRule :: (a -> a -> SeparatorRule) -> (SExpr a -> SExpr a -> SeparatorRule)
+sexprRule r =
+  let sepRule (SList _) _ = SOptional
+      sepRule _ (SList _) = SOptional
+      sepRule (SAtom x1) (SAtom x2) = r x1 x2
+  in sepRule
+
 
 -- | The 'parseSExpr' function return a parser for parsing
 -- S-expression ('SExpr'), that is either an atom (@'SAtom' _@) or a
 -- list @'SList' _ _@. See also 'decodeOne' and 'decode'.
-parseSExpr :: (MonadParsec e s m) =>
-              SListDefinition m a ->
-              m (SExpr a)
-parseSExpr (SListDefinition s e a sp r) =
-    let sepRule' (SList _) _ = SOptional
-        sepRule' _ (SList _) = SOptional
-        sepRule' (SAtom x1) (SAtom x2) = r x1 x2
+sexpr :: (MonadParsec e s m) =>
+             m b ->
+             m c ->
+             m a ->
+             m sep ->
+             (a -> a -> SeparatorRule) ->
+             m (SExpr a)
+sexpr s e a sep r =
+  let p = (SAtom <$> a) <|> (SList <$> sepEndByList (s >> optional sep) e p sep (sexprRule r))
+  in p
 
-        def' = SListDefinition s e p sp sepRule'
+manySExpr :: (MonadParsec e s m) =>
+             m b ->
+             m c ->
+             m a ->
+             m sep ->
+             (a -> a -> SeparatorRule) ->
+             m [SExpr a]
+manySExpr s e a sep r =
+  let rSExpr = sexprRule r
+      p = (SAtom <$> a) <|> (SList <$> sepEndByList (s >> optional sep) e p sep rSExpr)
+  in sepEndByRule p sep rSExpr
 
-        p = (SAtom <$> a) <|> (SList <$> sepByList def')
-    in p  
+manySExpr1 :: (MonadParsec e s m) =>
+             m b ->
+             m c ->
+             m a ->
+             m sep ->
+             (a -> a -> SeparatorRule) ->
+             m [SExpr a]
+manySExpr1 s e a sep r =
+  let rSExpr = sexprRule r
+      p = (SAtom <$> a) <|> (SList <$> sepEndByList (s >> optional sep) e p sep rSExpr)
+  in sepEndByRule1 p sep rSExpr
 
--- | The 'decodeOne' function return a parser for parsing a file
--- containing only one S-expression ('SExpr'). It can parse extra
--- whitespace at the beginning and at the end of the file. See also
--- 'parseSExpr' and 'decode'.
-decodeOne :: (MonadParsec e s m) => SListDefinition m a -> m (SExpr a)
-decodeOne def@(SListDefinition _ _ _ ws _) =
-  optional ws *> parseSExpr def <* (optional ws >> eof)
-
--- | The 'decode' function return a parser for parsing a file
--- containing many S-expression ('SExpr'). It can parse extra
--- whitespace at the beginning and at the end of the file. See also
--- 'parseSExpr' and 'decodeOne'.
-decode :: (MonadParsec e s m) => SListDefinition m a -> m [SExpr a]
-decode def@(SListDefinition _ _ _ ws spaceRule') =
-  let sepRule' (SList _) _ = SOptional
-      sepRule' _ (SList _) = SOptional
-      sepRule' (SAtom x1) (SAtom x2) = spaceRule' x1 x2
-  in optional ws *> sepEndByRule (parseSExpr def) ws sepRule' <* eof